camera-ready copy submitted to IEEE HST'13
It was necessary to modify the process to generate a DVI file and use
EPS versions of the figures because IEEE mandates use of their PDF
validator tool for acceptance, and the PDF file generated by PDFLaTeX
would not pass because not all the fonts were Type 1. Ironically, the
PDF their tool generated still contains one subset TrueType font, but it
passes their validator, so that's what got submitted.